Move generate_api_docs.py out of conf.py
